+++ ServerNode.cc
@@ -8,8 +8,8 @@ Define_Module(ServerNode);
 void ServerNode::initialize()
 {
     processing_delay=par("processingDelay").doubleValue();
-    queue_size=par("queueSize").longValue();
-    watts_per_second=par("wattsConsumePerSecond").longValue();
+    queue_size=par("queueSize").intValue();
+    watts_per_second=par("wattsConsumePerSecond").intValue();
